+++ packages/dashboard/src/components/Card/Chart/Chart.types.ts
@@ -103,6 +103,11 @@ export interface IChartProps {
    * One of the timeRange must be specified to line chart when Date type data is selected for line chart. The default is 180Days format
    */
   timeRange?: '7Days' | '30Days' | '90Days' | '180Days';
+
+  /**
+   * Ignore stack bar chart default style when data points count is less than two
+   */
+  ignoreStackBarChartDefaultStyle?: boolean;
 }
 
 export interface IChartInternalProps extends IChartProps {