+++ packages/office-ui-fabric-react/src/components/ContextualMenu/ContextualMenu.classNames.ts
@@ -4,6 +4,7 @@ import { ITheme, mergeStyleSets, getGlobalClassNames } from '../../Styling';
 import { IVerticalDividerClassNames } from '../Divider/VerticalDivider.types';
 import { memoizeFunction, IsFocusVisibleClassName } from '../../Utilities';
 import { IContextualMenuItemStyles, IContextualMenuItemStyleProps } from './ContextualMenuItem.types';
+import { IContextualMenuSubComponentStyles } from './ContextualMenu.types';
 
 /**
  * @deprecated in favor of mergeStyles API.
@@ -14,6 +15,7 @@ export interface IContextualMenuClassNames {
   list: string;
   header: string;
   title: string;
+  subComponentStyles: IContextualMenuSubComponentStyles;
 }
 
 /**