+++ packages/charting/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @uifabric/charting
 
-This log was last generated on Mon, 29 Oct 2018 18:04:30 GMT and should not be manually modified.
+This log was last generated on Thu, 01 Nov 2018 12:31:45 GMT and should not be manually modified.
+
+## 0.26.3
+Thu, 01 Nov 2018 12:31:45 GMT
+
+### Patches
+
+- added hovercard for horizontalBar chart
+
+## 0.26.2
+Wed, 31 Oct 2018 12:32:41 GMT
+
+### Patches
+
+- Enable ignore fixed stackbarchart style when data point is less than 2
 
 ## 0.26.1
 Mon, 29 Oct 2018 18:04:30 GMT